# src/main/resources/schema/swagger/covid19-swagger-spec.yml
swagger: '2.0'
info:
  version: 0.0.1
  title:  Covid19-case-follow-up
  description: ''
host: 'localhost:8080'
basePath: /
securityDefinitions:
  basicAuth:
    type: basic
schemes:
  - http
...
paths:
...
  '/v1/facility':
      put:
        operationId: registerNewFacility
        summary: 'Register new facility'
        tags:
          - HealthFacility
        consumes:
          - application/json
        parameters:
          - name: facilityData
            in: body
            schema:
              $ref: '#/definitions/request-save-facility'
        responses:
          '200':
            description: ''
            schema:
              $ref: '#/definitions/response-base'
          '401':
            $ref: '#/responses/response-unauthorized'
          '500':
            $ref: '#/responses/response-internal-server-error-500'
...
definitions:
  response-base:
    type: object
    properties:
      success:
        type: boolean
      resultCode:
        type: integer
        format: int32
      message:
        type: string
      errors:
        type: array
        items:
          type: string
      transactionId:
        type: string
      type:
        type: string
    discriminator: type
    required:
      - success
      - resultCode
...
  model-questionnier:
    type: object
    properties:
      id:
        type: integer
        format: int32
      question:
        type: string
      options:
        description: Should be a valid json data
        type: string
      category:
        $ref: '#/definitions/model-enum-id-value'
      parentId:
        type: integer
        format: int32
      description:
        type: string
      modifiedBy:
        type: integer
        format: int32
      insertDate:
        type: string
        format: date-time
      modifiedDate:
        type: string
        format: date-time
    required:
      - question
      - options
      - category
  model-attribute:
    type: object
    properties:
      attName:
        type: string
      attType:
        type: string
      attCode:
        type: string
      attLabel:
        type: string
      attDesc:
        type: string
  model-pui-follow-up:
    type: object
    properties:
      puiCaseCode:
        type: string
      qId:
        type: integer
        format: int32
      question:
        type: string
      selectedOption:
        type: string
      description:
        type: string
      modifiedBy:
        type: integer
        format: int32
      insertDate:
        type: string
        format: date-time
    required:
      - puiCaseCode
      - qId
      - selectedOption
  model-enum-id-value:
    type: object
    properties:
      id:
        type: integer
        format: int32
      value:
        type: string
  model-questionnier-list:
    type: object
    properties:
      cases:
        type: array
        items:
          $ref: '#/definitions/model-questionnier'
  model-case-list:
    type: object
    properties:
      cases:
        type: array
        items:
          $ref: '#/definitions/model-case'
  model-attribute-list:
    type: object
    properties:
      attributes:
        type: array
        items:
          $ref: '#/definitions/model-attribute'
  model-pui-follow-up-list:
    type: object
    properties:
      list:
        type: array
        items:
          $ref: '#/definitions/model-pui-follow-up'
    required:
      - list
  request-save-case:
    allOf:
      - $ref: '#/definitions/model-case'
...
  request-save-facility:
    allOf:
      required:
        - list
    request-save-case:
      allOf:
        - $ref: '#/definitions/model-case'
    request-save-questionnier:
      allOf:
        - $ref: '#/definitions/model-questionnier'
    request-save-follow-up:
      allOf:
        - $ref: '#/definitions/request-save-case'
        - $ref: '#/definitions/model-pui-follow-up-list'
    response-case-single:
      allOf:
        - $ref: '#/definitions/response-base'
        - type: object
          properties:
            returnValue:
              $ref: '#/definitions/model-case'
    response-case-list:
      allOf:
        - $ref: '#/definitions/response-base'
        - type: object
          properties:
            returnValue:
              $ref: '#/definitions/model-case-list'
    response-questionnier-single:
      allOf:
        - $ref: '#/definitions/response-base'
        - type: object
          properties:
            returnValue:
              $ref: '#/definitions/model-questionnier'
    response-questionnier-list:
      allOf:
        - $ref: '#/definitions/response-base'
        - type: object
          properties:
            returnValue:
              $ref: '#/definitions/model-questionnier-list'
    response-attribute:
      allOf:
        - $ref: '#/definitions/response-base'
        - type: object
          properties:
            returnValue:
              $ref: '#/definitions/model-attribute'
    response-attributes-list:
      allOf:
        - $ref: '#/definitions/response-base'
        - type: object
          properties:
            returnValue:
              $ref: '#/definitions/model-attribute-list'
    response-pui-follow-up-single:
      allOf:
        - $ref: '#/definitions/response-base'
        - type: object
          properties:
            returnValue:
              $ref: '#/definitions/model-pui-follow-up'
    response-pui-follow-up-list:
      allOf:
        - $ref: '#/definitions/response-base'
        - type: object
          properties:
            returnValue:
              $ref: '#/definitions/model-pui-follow-up-list'
    request-save-facility:
      allOf:
      - $ref: '#/definitions/model-health-facility'
responses:
  response-internal-server-error-500:
    description: Internal Server error
    schema:
      $ref: '#/definitions/response-base'
  response-unauthorized:
    description: Authentication information is missing or invalid
    schema:
      $ref: '#/definitions/response-base'
    headers:
      WWW_Authenticate:
        type: string